pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-access"
version = "0.4.0"
description = "Enable/disable agent SSH and GitHub collaborator access from a YAML project config"
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
requires-python = ">=3.10"
authors = [
  { name = "agent-access contributors" },
]
keywords = ["ssh", "github", "collaborator", "authorized_keys", "agent", "automation", "cli"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "Intended Audience :: System Administrators",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: System :: Systems Administration",
  "Topic :: Utilities",
]
dependencies = [
  "pyyaml>=6.0",
  "requests>=2.31.0",
  "paramiko>=3.4.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "build>=1.0",
]

[project.urls]
Repository = "https://github.com/ajlyakhov/agent-access"

[project.scripts]
agent-access = "agent_access.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["agent_access*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-q"
